refactoring module names and refering
