+++ t_smiles/mol_utils/rdk_utils/frag/rdk_frag_brics.py
@@ -1,7 +1,7 @@
 import rdkit.Chem as Chem
 from rdkit.Chem import BRICS
 
-from MolUtils.RDKUtils.Frag.RDKFragUtil import RDKFragUtil
+from t_smiles.mol_utils.rdk_utils.frag.rdk_frag_util import RDKFragUtil
 
 class RDKFragBrics:
