+++ t_smiles/tools/math_utils.py
